+++ HISTORY.md
@@ -3,3 +3,8 @@
 
 ## Version 1.0.8
   - renamed to dbrequests
+
+## Version 1.1
+  - added handling of percentage signs
+  - added possibility of automated comment removal
+  - ensured the compatibility of raw SQL code with code sent via dbrequests